+++ Rules/IPv4/Inbound/Broadcast.ps1
@@ -51,35 +51,64 @@ $BroadcastAddress = Get-Broadcast
 Update-Context "IPv$IPVersion" $Direction $Group @Logs
 if (!(Approve-Execute @Logs)) { exit }
 
-# First remove all existing rules matching group
+# First remove all existing rules matching grou-Group $Group p
 Remove-NetFirewallRule -PolicyStore $PolicyStore -Group $Group -Direction $Direction -ErrorAction Ignore @Logs
 
 #
 # TODO: currently handling only UDP, also broadcast falls into multicast space
 #
 
-New-NetFirewallRule -Platform $Platform `
-	-DisplayName "Limited Broadcast" -Service Any -Program System `
-	-PolicyStore $PolicyStore -Enabled True -Action Block -Group $Group -Profile $FirewallProfile -InterfaceType $Interface `
-	-Direction $Direction -Protocol UDP -LocalAddress $LimitedBroadcast -RemoteAddress LocalSubnet4 -LocalPort Any -RemotePort Any `
-	-EdgeTraversalPolicy Block -LocalUser $NT_AUTHORITY_System -LocalOnlyMapping $false -LooseSourceMapping $false `
-	-Description "" @Logs | Format-Output @Logs
+# NOTE: Limited broadcast can be used by DHCP
+New-NetFirewallRule -DisplayName "Limited Broadcast" `
+	-Platform $Platform -PolicyStore $PolicyStore -Profile Any `
+	-Service Any -Program System -Group $Group `
+	-Enabled True -Action Allow -Direction $Direction -Protocol UDP `
+	-LocalAddress $LimitedBroadcast -RemoteAddress Any `
+	-LocalPort Any -RemotePort Any `
+	-LocalUser $NT_AUTHORITY_System -EdgeTraversalPolicy Block `
+	-InterfaceType $Interface `
+	-LocalOnlyMapping $false -LooseSourceMapping $false `
+	-Description "" `
+	@Logs | Format-Output @Logs
 
-New-NetFirewallRule -Platform $Platform `
-	-DisplayName "LAN Broadcast" -Service Any -Program System `
-	-PolicyStore $PolicyStore -Enabled True -Action Allow -Group $Group -Profile $FirewallProfile -InterfaceType $Interface `
-	-Direction $Direction -Protocol UDP -LocalAddress $BroadcastAddress -RemoteAddress LocalSubnet4 -LocalPort Any -RemotePort Any `
-	-EdgeTraversalPolicy Block -LocalUser $NT_AUTHORITY_System -LocalOnlyMapping $false -LooseSourceMapping $false `
-	-Description "" @Logs | Format-Output @Logs
+# TODO: unsure if Intranet makes sense here
+New-NetFirewallRule -DisplayName "Limited Broadcast" `
+	-Platform $Platform -PolicyStore $PolicyStore -Profile Public `
+	-Service Any -Program System -Group $Group `
+	-Enabled True -Action Block -Direction $Direction -Protocol UDP `
+	-LocalAddress $LimitedBroadcast -RemoteAddress LocalSubnet4, Intranet4 `
+	-LocalPort Any -RemotePort Any `
+	-LocalUser $NT_AUTHORITY_System -EdgeTraversalPolicy Block `
+	-InterfaceType $Interface `
+	-LocalOnlyMapping $false -LooseSourceMapping $false `
+	-Description "Explicitly deny broadcast traffic on public subnets" `
+	@Logs | Format-Output @Logs
+
+New-NetFirewallRule -DisplayName "LAN Broadcast" `
+	-Platform $Platform -PolicyStore $PolicyStore -Profile $FirewallProfile `
+	-Service Any -Program System -Group $Group `
+	-Enabled True -Action Allow -Direction $Direction -Protocol UDP `
+	-LocalAddress $BroadcastAddress -RemoteAddress LocalSubnet4 `
+	-LocalPort Any -RemotePort Any `
+	-LocalUser $NT_AUTHORITY_System -EdgeTraversalPolicy Block `
+	-InterfaceType $Interface `
+	-LocalOnlyMapping $false -LooseSourceMapping $false `
+	-Description "" `
+	@Logs | Format-Output @Logs
 
 # TODO: check if virtual adapter exists and apply rule
 
-<# New-NetFirewallRule -Platform $Platform `
--DisplayName "Microsoft Wireless WiFi adapter" -Service Any -Program System `
--PolicyStore $PolicyStore -Enabled True -Action Allow -Group $Group -Profile $FirewallProfile -InterfaceType $Interface `
--Direction $Direction -Protocol UDP -LocalAddress Any -RemoteAddress 192.168.137.255 -LocalPort Any -RemotePort Any `
--LocalUser $NT_AUTHORITY_System -LocalOnlyMapping $false -LooseSourceMapping $false `
--Description "" @Logs | Format-Output @Logs
+<# New-NetFirewallRule -DisplayName "Microsoft Wireless WiFi adapter" `
+-Platform $Platform -PolicyStore $PolicyStore -Profile $FirewallProfile `
+-Service Any -Program System -Group $Group `
+-Enabled True -Action Allow -Direction $Direction -Protocol UDP `
+-LocalAddress Any -RemoteAddress 192.168.137.255 `
+-LocalPort Any -RemotePort Any `
+-LocalUser $NT_AUTHORITY_System -EdgeTraversalPolicy Block `
+-InterfaceType $Interface `
+-LocalOnlyMapping $false -LooseSourceMapping $false `
+-Description "" `
+@Logs | Format-Output @Logs
  #>
 
 Update-Logs